Add tests for project and task
