+++ test/task.test.js
@@ -1,0 +1,10 @@
+import Task from '../src/task'
+
+test('create task with valid inputs', () => {
+  const task = new Task('test', '05-05-2020', 'low');
+  expect(task.name).toBe('test');
+  expect(task.date).toBe('05-05-2020');
+  expect(task.priority).toBe('low');
+});
+
+
